This is, I hope, the last major revision of the code. Everything was flattened to one package, so if you relied on these classes, you may have to reorganize imports. The public API did not change much, so everything should still work.
Look at the TournamentRunner class and run it. It contains methods that run players originating from local (IPlayerFactory) or Corba (ICPlayerFactory) factories. These methods implement a single round timeout (very short, so that you can see it works) and re-run players for multiple rounds. This will be the basis of the pairwise competition, I'm sure.

The wrapper code for local players in ICPlayerControllerAdapter is not trivial. It runs an asynchronous event pump thread and silently (well, it logs them) drops events and frames that have been queued in between controller calls. It should work for your classes, if you have a local controller. Inspect the code, understand it. Try to find bugs, if you can.

I tend to believe that controller state callback MUST include the number of frames after which the signal is invalidated.

git-svn-id: https://ophelia.cs.put.poznan.pl/svn/prv-dweiss@3704 7e8c87aa-b7f8-0310-9a37-b181ab0e442b